refactor(users): change dir
